% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{conf_int}
\alias{conf_int}
\title{Confidence Interval}
\usage{
conf_int(value =  , se = , level = )
}
\arguments{
\item{value}{The numeric value of the statistic you would like a CI for.}

\item{se}{The Standard Error of the statistic you would like a CI for.}

\item{level}{The desired Confidence Level}
}
\description{
A function that produces an equal-tailed 95% CI for an inputed value.
}
\author{
David B
}
