% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exp_dist.R
\name{exp_density}
\alias{exp_density}
\title{The Exponetial Distribution}
\usage{
exp_density(size = ,mean= )
}
\arguments{
\item{size}{Specify the number of observations.}

\item{mean}{Specify the desired mean of the distribution.}
}
\description{
A neat little funtion that produces a exponetially distributed data
set.
}
\author{
David B
}
