\name{Exponential Distribution}
\alias{Exponential Distribution}

\title{
exp_density()
}

\usage{
Generate an Exponential Distribution.
}

\arguments{
  \item{size}{
Number of observations generated.
}
\item{mean}{
Specify the mean of the density.
}
}


\author{
David B.
}
\examples{
exp_density(1000, 5) generates an exponential distribution with
1000 observations and a mean of 5.
}

