% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessgame.R
\name{guess_game}
\alias{guess_game}
\title{Two Player Guessing Game}
\usage{
guess_game(max = 10 ,guess1 = 5 ,guess2 = 3)
}
\arguments{
\item{max}{Specificy the maximum possible value in the range of numbers
to be guessed.}

\item{guess1}{Player one's guess.}

\item{guess2}{Player two's guess.}
}
\description{
A two-player guessing game! Numbers that each player guess
range from 1 to Max.
}
\author{
David B
}
