% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme_black.R
\name{my_theme2}
\alias{my_theme2}
\title{Custom ggtheme 2}
\usage{
my_theme2()
}
\description{
Another custom ggtheme with a gnarly black background. I plan to add to this
in the future.
}
\author{
David B
}
