\name{MLE of Normal Data}
\alias{MLE of Normal Data}
\title{
normal_mle()
}
\description{
Calculate the maximum likelihood estimates for the mean and variance of normally distributed data. Outputs the mean then the variance.
}
\usage{
normal_mle()
}
\arguments{
  \item{mydata}{
The dataset to be used.
}
}

\author{
DB
}
