% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powercalc.R
\name{power_calc}
\alias{power_calc}
\title{Approximate Power Calculation}
\usage{
power_calc(samples= , alpha= , mu0= , mu1= )
}
\arguments{
\item{samples}{Sample size.}

\item{alpha}{Significance level.}

\item{mu0}{Mean under the null hypothesis.}

\item{mu1}{The "true" mean.}
}
\description{
An approximate power calculation.
}
\author{
David B
}
