% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{sample_size}
\alias{sample_size}
\title{Sample size calculator}
\usage{
sample_size(alpha = , power = , sd = , mu0 = , mu1 = )
}
\arguments{
\item{alpha}{Desired Significance level}

\item{sd}{Standard Deviation}

\item{mu0}{Mean under the null hypothesis}

\item{mu1}{Mean under the alternative hypothesis}

\item{power}{Desired Power}
}
\description{
Gives the approximate sample size needed for desired power,
significance level, estimated standard deviation, and mean under null and
alternative hypothesis.
}
