% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_search.R
\name{search_str}
\alias{search_str}
\title{String Search}
\usage{
search_str(text = "Search for string", string = "string")
}
\arguments{
\item{text}{The text to perform search on.}

\item{string}{The string you want to search for in the text. Must be inside " ".}
}
\description{
Search for a specific string in text.
}
