% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{time_series}
\alias{time_series}
\title{Simple Time Series}
\usage{
time_series(mydata= , p1= , d1= , q1=)
}
\arguments{
\item{mydata}{Specify the data to be used.}

\item{p1}{Specify the number of autoregressive terms.}

\item{d1}{Specify the degree of differencing.}

\item{q1}{Specify the number of moving average terms.}
}
\description{
Specify an ARMA(p,d,q) to fit data and produce a forecast.
}
\author{
David B
}
