% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topfreq.R
\name{top_n}
\alias{top_n}
\title{Top n Frequencies}
\usage{
top_n(mydata= , groupings= " ", n= 1)
}
\arguments{
\item{mydata}{Specify the data you will use.}

\item{groupings}{Specify the variable(s) you will group by.}

\item{n}{Specify the top n groups to display.}
}
\description{
Display the top n groups based on frequency.
}
\author{
David B
}
