\name{exp_density}
\alias{exp_density}

\title{
Generate an Exponetial Distribution
}

\description{
Generate an Exponetioal Dist. with a specific number of obs. and mean.
}

\usage{
exp_density(size, mean)
}

\arguments{
  \item{size}{
Number of observations generated.
}
\item{mean}{
Specify the mean of the density.
}
}


\author{
David B.
}
\examples{
exp_density(1000, 5) generates an exponential distribution with
1000 observations and a mean of 5.
}

