% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{normal_mle}
\alias{normal_mle}
\title{MLE of Normal Distribution}
\usage{
normal_mle(mydata= )
}
\arguments{
\item{mydata}{The data you would like to estimate the parameters with.}

\item{var}{This is the column number of the variable of interest in your data.fram}
}
\description{
Calculate the Maximum Likelihood Estimates (MLEs) for the mean and
variance of for a normal distriubtion.
}
\author{
David B
}
