\name{normal_mle}
\alias{normal_mle}
\title{
MLE of Normal Data
}
\description{
Calculate the maximum likelihood estimates for the mean and variance of normally distributed data. Outputs the mean then the variance.
}
\usage{
normal_mle(mydata)
}
\arguments{
  \item{mydata}{
The dataset to be used.
}
}

\author{
DB
}
